[XABT] Split up the `<GeneratePackageManagerJava>` task.
