+++ src/Xamarin.Android.Build.Tasks/Tasks/GenerateCompressedAssembliesNativeSourceFiles.cs
@@ -101,7 +101,7 @@ namespace Xamarin.Android.Tasks
 
 					using (var sw = MemoryStreamPool.Shared.CreateStreamWriter ()) {
 						try {
-							composer.Generate (compressedAssemblies, GeneratePackageManagerJava.GetAndroidTargetArchForAbi (abi), sw, llvmIrFilePath);
+							composer.Generate (compressedAssemblies, GenerateNativeApplicationConfigAssemblies.GetAndroidTargetArchForAbi (abi), sw, llvmIrFilePath);
 						} catch {
 							throw;
 						} finally {
